Remove useless duplication
